+++ CHANGELOG.md
@@ -8,6 +8,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Failed to parse tool specifications in foreman.toml if they were an inline table ([#36])
+
+[#36]: https://github.com/rojo-rbx/rokit/pull/36
+
 ## `0.1.4` - July 11th, 2024
 
 ### Fixed